+++ src/StateSmith/Output/Gil/C99/ICFileNamer.cs
@@ -1,0 +1,9 @@
+#nullable enable
+
+namespace StateSmith.Output.Gil.C99;
+
+public interface ICFileNamer
+{
+    string MakeHFileName();
+    string MakeCFileName();
+}