+++ src/StateSmithTest/C99BalancedCoder1/CBuilderTest.cs
@@ -1,5 +1,4 @@
 //todo_low rework to test new Balanced1 code
-//using StateSmith.Output.C99BalancedCoder1;
 //using Xunit;
 //using System.Linq;
 //using StateSmith.SmGraph;