+++ src/StateSmithTest/C99BalancedCoder1/CEnumBuilderTest.cs
@@ -2,7 +2,6 @@
 
 //using FluentAssertions;
 //using StateSmith.SmGraph;
-//using StateSmith.Output.C99BalancedCoder1;
 //using StateSmith.Runner;
 //using System;
 //using System.Collections.Generic;