+++ src/StateSmithTest/C99BalancedCoder1/CHeaderBuilderTest.cs
@@ -1,6 +1,5 @@
 //todo_low rework to test new Balanced1 code
 
-//using StateSmith.Output.C99BalancedCoder1;
 //using Xunit;
 
 //namespace StateSmithTest.C99BalancedCoder1