+++ src/StateSmithTest/C99BalancedCoder1/ExamplesTestHelpers.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
 using StateSmith.SmGraph;
-using StateSmith.Output.C99BalancedCoder1;
 using StateSmith.Runner;
 
 namespace StateSmithTest