+++ src/StateSmithTest/spec/1b/c/Spec1bCTests.cs
@@ -1,7 +1,6 @@
 using Spec.Spec1b;
 using StateSmith.Input.Expansions;
 using StateSmith.Output;
-using StateSmith.Output.C99BalancedCoder1;
 using StateSmith.Output.UserConfig;
 using StateSmith.Runner;
 using System;